Add type annotation for Blurhash component (#24750)
